+++ Cargo.toml
@@ -2,8 +2,10 @@
 members = [
     "data/binary",
     "data/catalog",
-    "data/definition_operations",
     "data/schema_planner",
+    "data_definition/operations",
+    "data_definition/execution_plan",
+    "data_manipulation/query_plan",
     "deprecated/ast",
     "deprecated/binder_deprecated",
     "deprecated/catalog_deprecated/data_manager",
@@ -25,6 +27,7 @@ members = [
     "annotated_tree",
     "query_parsing/parser",
     "query_parsing/sql-ast",
+    "query_validation/type_inference",
     "query_validation/validator",
     "entities/repr",
     "server/connection",
