+++ annotated_tree/src/lib.rs
@@ -12,84 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use definition::{FullTableName, SchemaName};
 use expr_operators::{DynamicItem, Operation, StaticItem};
 use types::{SqlFamilyType, SqlType};
-
-pub type AnalysisResult<A> = Result<A, AnalysisError>;
-
-#[derive(Debug, PartialEq)]
-pub struct TableInfo {
-    pub schema_name: String,
-    pub table_name: String,
-}
-
-impl TableInfo {
-    pub fn new<S: ToString, T: ToString>(schema_name: &S, table_name: &T) -> TableInfo {
-        TableInfo {
-            schema_name: schema_name.to_string(),
-            table_name: table_name.to_string(),
-        }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub struct CreateSchemaQuery {
-    pub schema_name: SchemaName,
-    pub if_not_exists: bool,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct DropSchemasQuery {
-    pub schema_names: Vec<SchemaName>,
-    pub cascade: bool,
-    pub if_exists: bool,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct ColumnInfo {
-    pub name: String,
-    pub sql_type: SqlType,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct CreateTableQuery {
-    pub table_info: TableInfo,
-    pub column_defs: Vec<ColumnInfo>,
-    pub if_not_exists: bool,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct DropTablesQuery {
-    pub table_infos: Vec<TableInfo>,
-    pub cascade: bool,
-    pub if_exists: bool,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct InsertQuery {
-    pub full_table_name: FullTableName,
-    pub column_types: Vec<SqlType>,
-    pub values: Vec<Vec<StaticEvaluationTree>>,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct ColumnDesc {
-    pub name: String,
-    pub sql_type: SqlType,
-    pub ord_num: usize,
-}
-
-impl From<(String, SqlType, usize)> for ColumnDesc {
-    fn from(tuple: (String, SqlType, usize)) -> ColumnDesc {
-        let (name, sql_type, ord_num) = tuple;
-        ColumnDesc {
-            name,
-            sql_type,
-            ord_num,
-        }
-    }
-}
 
 #[derive(Debug, PartialEq)]
 pub enum StaticEvaluationTree {
@@ -112,13 +36,6 @@ impl StaticEvaluationTree {
 }
 
 #[derive(Debug, PartialEq)]
-pub struct UpdateQuery {
-    pub full_table_name: FullTableName,
-    pub sql_types: Vec<SqlType>,
-    pub assignments: Vec<DynamicEvaluationTree>,
-}
-
-#[derive(Debug, PartialEq)]
 pub enum DynamicEvaluationTree {
     Operation {
         left: Box<DynamicEvaluationTree>,
@@ -126,123 +43,6 @@ pub enum DynamicEvaluationTree {
         right: Box<DynamicEvaluationTree>,
     },
     Item(DynamicItem),
-}
-
-#[derive(Debug, PartialEq)]
-pub struct SelectQuery {
-    pub full_table_name: FullTableName,
-    pub projection_items: Vec<DynamicEvaluationTree>,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct DeleteQuery {
-    pub full_table_name: FullTableName,
-}
-
-#[derive(Debug, PartialEq)]
-pub enum SchemaChange {
-    CreateSchema(CreateSchemaQuery),
-    DropSchemas(DropSchemasQuery),
-    CreateTable(CreateTableQuery),
-    DropTables(DropTablesQuery),
-}
-
-#[derive(Debug, PartialEq)]
-pub enum Write {
-    Insert(InsertQuery),
-    Update(UpdateQuery),
-    Delete(DeleteQuery),
-}
-
-#[derive(Debug, PartialEq)]
-pub enum QueryAnalysis {
-    DataDefinition(SchemaChange),
-    Write(Write),
-    Read(SelectQuery),
-}
-
-#[derive(Debug, PartialEq)]
-pub enum AnalysisError {
-    SchemaNamingError(String),
-    SchemaDoesNotExist(String),
-    SchemaAlreadyExists(String),
-    TableNamingError(String),
-    TableDoesNotExist(String),
-    TableAlreadyExists(String),
-    TypeIsNotSupported(String),
-    SyntaxError(String),
-    ColumnNotFound(String),
-    ColumnCantBeReferenced(String),                                  // Error code: 42703
-    InvalidInputSyntaxForType { sql_type: SqlType, value: String },  // Error code: 22P02
-    StringDataRightTruncation(SqlType),                              // Error code: 22001
-    DatatypeMismatch { column_type: SqlType, source_type: SqlType }, // Error code: 42804
-    AmbiguousFunction(Operation),                                    // Error code: 42725
-    UndefinedFunction(Operation),                                    // Error code: 42883
-    FeatureNotSupported(Feature),
-}
-
-impl AnalysisError {
-    pub fn schema_naming_error<M: ToString>(message: M) -> AnalysisError {
-        AnalysisError::SchemaNamingError(message.to_string())
-    }
-
-    pub fn schema_does_not_exist<S: ToString>(schema_name: S) -> AnalysisError {
-        AnalysisError::SchemaDoesNotExist(schema_name.to_string())
-    }
-
-    pub fn schema_already_exists<S: ToString>(schema_name: S) -> AnalysisError {
-        AnalysisError::SchemaAlreadyExists(schema_name.to_string())
-    }
-
-    pub fn table_naming_error<M: ToString>(message: M) -> AnalysisError {
-        AnalysisError::TableNamingError(message.to_string())
-    }
-
-    pub fn table_does_not_exist<T: ToString>(table_name: T) -> AnalysisError {
-        AnalysisError::TableDoesNotExist(table_name.to_string())
-    }
-
-    pub fn table_already_exists<T: ToString>(table_name: T) -> AnalysisError {
-        AnalysisError::TableAlreadyExists(table_name.to_string())
-    }
-
-    pub fn type_is_not_supported<T: ToString>(type_name: T) -> AnalysisError {
-        AnalysisError::TypeIsNotSupported(type_name.to_string())
-    }
-
-    pub fn syntax_error(message: String) -> AnalysisError {
-        AnalysisError::SyntaxError(message)
-    }
-
-    pub fn column_not_found<C: ToString>(column_name: C) -> AnalysisError {
-        AnalysisError::ColumnNotFound(column_name.to_string())
-    }
-
-    pub fn column_cant_be_referenced<C: ToString>(column_name: C) -> AnalysisError {
-        AnalysisError::ColumnCantBeReferenced(column_name.to_string())
-    }
-
-    pub fn invalid_input_syntax_for_type<V: ToString>(sql_type: SqlType, value: V) -> AnalysisError {
-        AnalysisError::InvalidInputSyntaxForType {
-            sql_type,
-            value: value.to_string(),
-        }
-    }
-
-    pub fn string_data_right_truncation(sql_type: SqlType) -> AnalysisError {
-        AnalysisError::StringDataRightTruncation(sql_type)
-    }
-
-    pub fn datatype_mismatch(column_type: SqlType, source_type: SqlType) -> AnalysisError {
-        AnalysisError::DatatypeMismatch {
-            column_type,
-            source_type,
-        }
-    }
-
-    pub fn feature_not_supported(feature: Feature) -> AnalysisError {
-        AnalysisError::FeatureNotSupported(feature)
-    }
 }
 
 #[derive(Debug, PartialEq)]
