+++ data_definition/execution_plan/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "data_definition_execution_plan"
+version = "0.1.0"
+authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
+edition = "2018"
+
+[dependencies]
+definition = { path = "../../entities/definition" }
+types = { path = "../../entities/types" }
