+++ data_definition/execution_plan/src/lib.rs
@@ -1,0 +1,90 @@
+// Copyright 2020 - present Alex Dukhno
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use definition::SchemaName;
+use types::SqlType;
+
+#[derive(Debug, PartialEq)]
+pub struct TableInfo {
+    pub schema_name: String,
+    pub table_name: String,
+}
+
+impl TableInfo {
+    pub fn new<S: ToString, T: ToString>(schema_name: &S, table_name: &T) -> TableInfo {
+        TableInfo {
+            schema_name: schema_name.to_string(),
+            table_name: table_name.to_string(),
+        }
+    }
+}
+
+#[derive(Debug, PartialEq)]
+pub struct CreateSchemaQuery {
+    pub schema_name: SchemaName,
+    pub if_not_exists: bool,
+}
+
+#[derive(Debug, PartialEq)]
+pub struct DropSchemasQuery {
+    pub schema_names: Vec<SchemaName>,
+    pub cascade: bool,
+    pub if_exists: bool,
+}
+
+#[derive(Debug, PartialEq)]
+pub struct ColumnInfo {
+    pub name: String,
+    pub sql_type: SqlType,
+}
+
+#[derive(Debug, PartialEq)]
+pub struct CreateTableQuery {
+    pub table_info: TableInfo,
+    pub column_defs: Vec<ColumnInfo>,
+    pub if_not_exists: bool,
+}
+
+#[derive(Debug, PartialEq)]
+pub struct DropTablesQuery {
+    pub table_infos: Vec<TableInfo>,
+    pub cascade: bool,
+    pub if_exists: bool,
+}
+
+#[derive(Debug, PartialEq)]
+pub struct ColumnDesc {
+    pub name: String,
+    pub sql_type: SqlType,
+    pub ord_num: usize,
+}
+
+impl From<(String, SqlType, usize)> for ColumnDesc {
+    fn from(tuple: (String, SqlType, usize)) -> ColumnDesc {
+        let (name, sql_type, ord_num) = tuple;
+        ColumnDesc {
+            name,
+            sql_type,
+            ord_num,
+        }
+    }
+}
+
+#[derive(Debug, PartialEq)]
+pub enum SchemaChange {
+    CreateSchema(CreateSchemaQuery),
+    DropSchemas(DropSchemasQuery),
+    CreateTable(CreateTableQuery),
+    DropTables(DropTablesQuery),
+}
