+++ data_definition/operations/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "definition_operations"
+name = "data_definition_operations"
 version = "0.1.0"
 authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
 edition = "2018"
