+++ data_manipulation/query_plan/Cargo.toml
@@ -1,15 +1,10 @@
 [package]
-name = "schema_planner"
+name = "data_manipulation_query_plan"
 version = "0.1.0"
 authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
 edition = "2018"
-publish = false
 
 [dependencies]
 annotated_tree = { path = "../../annotated_tree" }
-data_definition_operations = { path = "../../data_definition/operations" }
-data_definition_execution_plan = { path = "../../data_definition/execution_plan" }
-
-[dev-dependencies]
 definition = { path = "../../entities/definition" }
 types = { path = "../../entities/types" }
