+++ query_analysis/query_analyzer/Cargo.toml
@@ -6,8 +6,10 @@ edition = "2018"
 publish = false
 
 [dependencies]
-analysis_tree = { path = "../../annotated_tree" }
+annotated_tree = { path = "../../annotated_tree" }
 catalog = { path = "../../data/catalog" }
+data_definition_execution_plan = { path = "../../data_definition/execution_plan" }
+data_manipulation_query_plan = { path = "../../data_manipulation/query_plan" }
 expr_operators = { path = "../expr_operators" }
 meta_def = { path = "../../meta_def" }
 definition = { path = "../../entities/definition" }
@@ -17,4 +19,4 @@ sql_ast = { path = "../../query_parsing/sql-ast" }
 bigdecimal = { version = "0.2.0", features = ["string-only"] }
 
 [dev-dependencies]
-definition_operations = { path = "../../data/definition_operations" }
+data_definition_operations = { path = "../../data_definition/operations" }
