+++ query_validation/type_inference/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "type_inference"
+version = "0.1.0"
+authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
+edition = "2018"
+
+[dependencies]
