+++ query_validation/validator/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 publish = false
 
 [dependencies]
-analysis_tree = { path = "../../annotated_tree" }
+annotated_tree = { path = "../../annotated_tree" }
 expr_operators = { path = "../../query_analysis/expr_operators" }
 types = { path = "../../entities/types" }
